refactor: pr comments for code conventions
