+++ packages/shared/src/lib/core/layer-2/actions/index.ts
@@ -1,4 +1,6 @@
+export * from './estimateGasForLayer1ToLayer2Transaction'
 export * from './fetchSelectedAccountLayer2Balance'
 export * from './generateAndStoreEvmAddressForAccount'
+export * from './getGasFeesForLayer1ToLayer2Transaction'
 export * from './getGasPriceInWei'
 export * from './pollLayer2Tokens'
