+++ packages/shared/src/lib/core/layer-2/utils/getLayer2MetadataForTransfer.ts
@@ -1,9 +1,9 @@
-import { encodeAddress, encodeAssetAllowance, encodeSmartContractParameters } from '../helpers'
-import BigInteger from 'big-integer'
 import type { SendFlowParameters, TokenSendFlowParameters } from '@core/wallet/types'
+import BigInteger from 'big-integer'
+import { estimateGasForLayer1ToLayer2Transaction } from '../actions/estimateGasForLayer1ToLayer2Transaction'
 import { SpecialStream } from '../classes'
 import { ACCOUNTS_CONTRACT, EXTERNALLY_OWNED_ACCOUNT, GAS_LIMIT_MULTIPLIER, TRANSFER_ALLOWANCE } from '../constants'
-import { estimateGasForLayer1ToLayer2Transaction } from './estimateGasForLayer1ToLayer2Transaction'
+import { encodeAddress, encodeAssetAllowance, encodeSmartContractParameters } from '../helpers'
 
 export async function getLayer2MetadataForTransfer(sendFlowParameters: SendFlowParameters): Promise<string> {
     const metadataStream = new SpecialStream()
