+++ packages/shared/src/lib/core/layer-2/utils/index.ts
@@ -5,7 +5,6 @@ export * from './buildUnwrapAssetTargetAddress'
 export * from './calculateEstimatedGasFeeFromTransactionData'
 export * from './calculateGasFeeInGlow'
 export * from './calculateMaxGasFeeFromTransactionData'
-export * from './estimateGasForLayer1ToLayer2Transaction'
 export * from './evmAddressToAgentID'
 export * from './getAbiForContractType'
 export * from './getAgentBalanceParameters'
